Struct aws_sdk_wafv2::client::fluent_builders::CreateRuleGroup

pub struct CreateRuleGroup<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateRuleGroup<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateRuleGroupOutput, SdkError<CreateRuleGroupError>> where
R::Policy: SmithyRetryPolicy<CreateRuleGroupInputOperationOutputAlias, CreateRuleGroupOutput, CreateRuleGroupError, CreateRuleGroupInputOperationRetryAlias>,

pub fn name(self, inp: impl Into<String>) -> Self

The name of the rule group. You cannot change the name of a rule group after you create it.

pub fn set_name(self, input: Option<String>) -> Self

pub fn scope(self, inp: Scope) -> Self

Specifies whether this is for an Amazon CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an Amazon API Gateway REST API, or an AppSync GraphQL API.

To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows:

CLI - Specify the Region when you use the CloudFront scope: --scope=CLOUDFRONT --region=us-east-1.

API and SDKs - For all calls, use the Region endpoint us-east-1.

pub fn set_scope(self, input: Option<Scope>) -> Self

pub fn capacity(self, inp: i64) -> Self

The web ACL capacity units (WCUs) required for this rule group.

When you create your own rule group, you define this, and you cannot change it after creation. When you add or modify the rules in a rule group, WAF enforces this limit. You can check the capacity for a set of rules using CheckCapacity.

WAF uses WCUs to calculate and control the operating resources that are used to run your rules, rule groups, and web ACLs. WAF calculates capacity differently for each rule type, to reflect the relative cost of each rule. Simple rules that cost little to run use fewer WCUs than more complex rules that use more processing power. Rule group capacity is fixed at creation, which helps users plan their web ACL WCU usage when they use a rule group. The WCU limit for web ACLs is 1,500.

pub fn set_capacity(self, input: Option<i64>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the rule group that helps with identification.

pub fn set_description(self, input: Option<String>) -> Self

pub fn rules(self, inp: impl Into<Rule>) -> Self

Appends an item to Rules.

To override the contents of this collection use set_rules.

The Rule statements used to identify the web requests that you want to allow, block, or count. Each rule includes one top-level statement that WAF uses to identify matching web requests, and parameters that govern how WAF handles them.

pub fn set_rules(self, input: Option<Vec<Rule>>) -> Self

pub fn visibility_config(self, inp: VisibilityConfig) -> Self

Defines and enables Amazon CloudWatch metrics and web request sample collection.

pub fn set_visibility_config(self, input: Option<VisibilityConfig>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

An array of key:value pairs to associate with the resource.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn custom_response_bodies(
self,
k: impl Into<String>,
v: impl Into<CustomResponseBody>
) -> Self

Adds a key-value pair to CustomResponseBodies.

To override the contents of this collection use set_custom_response_bodies.

A map of custom response keys and content bodies. When you create a rule with a block action, you can send a custom response to the web request. You define these for the rule group, and then use them in the rules that you define in the rule group.

For information about customizing web requests and responses, see Customizing web requests and responses in WAF in the WAF Developer Guide.

For information about the limits on count and size for custom request and response settings, see WAF quotas in the WAF Developer Guide.

pub fn set_custom_response_bodies(
self,
input: Option<HashMap<String, CustomResponseBody>>
) -> Self