Struct aws_sdk_wafv2::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSWAF_20190729.

This client allows ergonomic access to a AWSWAF_20190729-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_web_acl(&self) -> AssociateWebACL<C, M, R>

pub fn check_capacity(&self) -> CheckCapacity<C, M, R>

pub fn create_ip_set(&self) -> CreateIPSet<C, M, R>

pub fn create_regex_pattern_set(&self) -> CreateRegexPatternSet<C, M, R>

pub fn create_rule_group(&self) -> CreateRuleGroup<C, M, R>

pub fn create_web_acl(&self) -> CreateWebACL<C, M, R>

pub fn delete_firewall_manager_rule_groups(
&self
) -> DeleteFirewallManagerRuleGroups<C, M, R>

pub fn delete_ip_set(&self) -> DeleteIPSet<C, M, R>

pub fn delete_logging_configuration(
&self
) -> DeleteLoggingConfiguration<C, M, R>

pub fn delete_permission_policy(&self) -> DeletePermissionPolicy<C, M, R>

pub fn delete_regex_pattern_set(&self) -> DeleteRegexPatternSet<C, M, R>

pub fn delete_rule_group(&self) -> DeleteRuleGroup<C, M, R>

pub fn delete_web_acl(&self) -> DeleteWebACL<C, M, R>

pub fn describe_managed_rule_group(&self) -> DescribeManagedRuleGroup<C, M, R>

pub fn disassociate_web_acl(&self) -> DisassociateWebACL<C, M, R>

pub fn get_ip_set(&self) -> GetIPSet<C, M, R>

pub fn get_logging_configuration(&self) -> GetLoggingConfiguration<C, M, R>

pub fn get_managed_rule_set(&self) -> GetManagedRuleSet<C, M, R>

pub fn get_permission_policy(&self) -> GetPermissionPolicy<C, M, R>

pub fn get_rate_based_statement_managed_keys(
&self
) -> GetRateBasedStatementManagedKeys<C, M, R>

pub fn get_regex_pattern_set(&self) -> GetRegexPatternSet<C, M, R>

pub fn get_rule_group(&self) -> GetRuleGroup<C, M, R>

pub fn get_sampled_requests(&self) -> GetSampledRequests<C, M, R>

pub fn get_web_acl(&self) -> GetWebACL<C, M, R>

pub fn get_web_acl_for_resource(&self) -> GetWebACLForResource<C, M, R>

pub fn list_available_managed_rule_groups(
&self
) -> ListAvailableManagedRuleGroups<C, M, R>

pub fn list_available_managed_rule_group_versions(
&self
) -> ListAvailableManagedRuleGroupVersions<C, M, R>

pub fn list_ip_sets(&self) -> ListIPSets<C, M, R>

pub fn list_logging_configurations(&self) -> ListLoggingConfigurations<C, M, R>

pub fn list_managed_rule_sets(&self) -> ListManagedRuleSets<C, M, R>

pub fn list_regex_pattern_sets(&self) -> ListRegexPatternSets<C, M, R>

pub fn list_resources_for_web_acl(&self) -> ListResourcesForWebACL<C, M, R>

pub fn list_rule_groups(&self) -> ListRuleGroups<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_web_ac_ls(&self) -> ListWebACLs<C, M, R>

pub fn put_logging_configuration(&self) -> PutLoggingConfiguration<C, M, R>

pub fn put_managed_rule_set_versions(
&self
) -> PutManagedRuleSetVersions<C, M, R>

pub fn put_permission_policy(&self) -> PutPermissionPolicy<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_ip_set(&self) -> UpdateIPSet<C, M, R>

pub fn update_managed_rule_set_version_expiry_date(
&self
) -> UpdateManagedRuleSetVersionExpiryDate<C, M, R>

pub fn update_regex_pattern_set(&self) -> UpdateRegexPatternSet<C, M, R>

pub fn update_rule_group(&self) -> UpdateRuleGroup<C, M, R>

pub fn update_web_acl(&self) -> UpdateWebACL<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self