Module aws_sdk_wafv2::error

Modules

waf_associated_item_exception

See WafAssociatedItemException

waf_duplicate_item_exception

See WafDuplicateItemException

waf_expired_managed_rule_group_version_exception

See WafExpiredManagedRuleGroupVersionException

waf_internal_error_exception

See WafInternalErrorException

waf_invalid_operation_exception

See WafInvalidOperationException

waf_invalid_parameter_exception

See WafInvalidParameterException

waf_invalid_permission_policy_exception

See WafInvalidPermissionPolicyException

waf_invalid_resource_exception

See WafInvalidResourceException

waf_limits_exceeded_exception

See WafLimitsExceededException

waf_nonexistent_item_exception

See WafNonexistentItemException

waf_optimistic_lock_exception

See WafOptimisticLockException

waf_service_linked_role_error_exception

See WafServiceLinkedRoleErrorException

waf_subscription_not_found_exception

See WafSubscriptionNotFoundException

waf_tag_operation_exception

See WafTagOperationException

waf_tag_operation_internal_error_exception

See WafTagOperationInternalErrorException

waf_unavailable_entity_exception

See WafUnavailableEntityException

Structs

AssociateWebACLError

CheckCapacityError

CreateIPSetError

CreateRegexPatternSetError

CreateRuleGroupError

CreateWebACLError

DeleteFirewallManagerRuleGroupsError

DeleteIPSetError

DeleteLoggingConfigurationError

DeletePermissionPolicyError

DeleteRegexPatternSetError

DeleteRuleGroupError

DeleteWebACLError

DescribeManagedRuleGroupError

DisassociateWebACLError

GetIPSetError

GetLoggingConfigurationError

GetManagedRuleSetError

GetPermissionPolicyError

GetRateBasedStatementManagedKeysError

GetRegexPatternSetError

GetRuleGroupError

GetSampledRequestsError

GetWebACLError

GetWebACLForResourceError

ListAvailableManagedRuleGroupVersionsError

ListAvailableManagedRuleGroupsError

ListIPSetsError

ListLoggingConfigurationsError

ListManagedRuleSetsError

ListRegexPatternSetsError

ListResourcesForWebACLError

ListRuleGroupsError

ListTagsForResourceError

ListWebACLsError

PutLoggingConfigurationError

PutManagedRuleSetVersionsError

PutPermissionPolicyError

TagResourceError

UntagResourceError

UpdateIPSetError

UpdateManagedRuleSetVersionExpiryDateError

UpdateRegexPatternSetError

UpdateRuleGroupError

UpdateWebACLError

WafAssociatedItemException

WAF couldn’t perform the operation because your resource is being used by another

WafDuplicateItemException

WAF couldn’t perform the operation because the resource that you tried to save is

WafExpiredManagedRuleGroupVersionException

The operation failed because the specified version for the managed rule group has

WafInternalErrorException

Your request is valid, but WAF couldn’t perform the operation because of a system

WafInvalidOperationException

The operation isn't valid.

WafInvalidParameterException

The operation failed because WAF didn't recognize a parameter in the request. For

WafInvalidPermissionPolicyException

The operation failed because the specified policy isn't in the proper format.

WafInvalidResourceException

WAF couldn’t perform the operation because the resource that you requested isn’t

WafLimitsExceededException

WAF couldn’t perform the operation because you exceeded your resource limit. For

WafNonexistentItemException

WAF couldn’t perform the operation because your resource doesn’t exist.

WafOptimisticLockException

WAF couldn’t save your changes because you tried to update or delete a resource

WafServiceLinkedRoleErrorException

WAF is not able to access the service linked role. This can be caused by a

WafSubscriptionNotFoundException

You tried to use a managed rule group that's available by subscription, but you aren't

WafTagOperationException

An error occurred during the tagging operation. Retry your request.

WafTagOperationInternalErrorException

WAF couldn’t perform your tagging operation because of an internal error. Retry

WafUnavailableEntityException

WAF couldn’t retrieve the resource that you requested. Retry your request.

Enums

AssociateWebACLErrorKind

CheckCapacityErrorKind

CreateIPSetErrorKind

CreateRegexPatternSetErrorKind

CreateRuleGroupErrorKind

CreateWebACLErrorKind

DeleteFirewallManagerRuleGroupsErrorKind

DeleteIPSetErrorKind

DeleteLoggingConfigurationErrorKind

DeletePermissionPolicyErrorKind

DeleteRegexPatternSetErrorKind

DeleteRuleGroupErrorKind

DeleteWebACLErrorKind

DescribeManagedRuleGroupErrorKind

DisassociateWebACLErrorKind

GetIPSetErrorKind

GetLoggingConfigurationErrorKind

GetManagedRuleSetErrorKind

GetPermissionPolicyErrorKind

GetRateBasedStatementManagedKeysErrorKind

GetRegexPatternSetErrorKind

GetRuleGroupErrorKind

GetSampledRequestsErrorKind

GetWebACLErrorKind

GetWebACLForResourceErrorKind

ListAvailableManagedRuleGroupVersionsErrorKind

ListAvailableManagedRuleGroupsErrorKind

ListIPSetsErrorKind

ListLoggingConfigurationsErrorKind

ListManagedRuleSetsErrorKind

ListRegexPatternSetsErrorKind

ListResourcesForWebACLErrorKind

ListRuleGroupsErrorKind

ListTagsForResourceErrorKind

ListWebACLsErrorKind

PutLoggingConfigurationErrorKind

PutManagedRuleSetVersionsErrorKind

PutPermissionPolicyErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateIPSetErrorKind

UpdateManagedRuleSetVersionExpiryDateErrorKind

UpdateRegexPatternSetErrorKind

UpdateRuleGroupErrorKind

UpdateWebACLErrorKind