Module aws_sdk_wafv2::model

Modules

action_condition

See ActionCondition

all

See All

all_query_arguments

See AllQueryArguments

allow_action

See AllowAction

and_statement

See AndStatement

block_action

See BlockAction

body

See Body

byte_match_statement

See ByteMatchStatement

condition

See Condition

count_action

See CountAction

custom_http_header

See CustomHttpHeader

custom_request_handling

See CustomRequestHandling

custom_response

See CustomResponse

custom_response_body

See CustomResponseBody

default_action

See DefaultAction

excluded_rule

See ExcludedRule

field_to_match

See FieldToMatch

filter

See Filter

firewall_manager_rule_group

See FirewallManagerRuleGroup

firewall_manager_statement

See FirewallManagerStatement

forwarded_ip_config

See ForwardedIpConfig

geo_match_statement

See GeoMatchStatement

http_header

See HttpHeader

http_request

See HttpRequest

ip_set

See IpSet

ip_set_forwarded_ip_config

See IpSetForwardedIpConfig

ip_set_reference_statement

See IpSetReferenceStatement

ip_set_summary

See IpSetSummary

json_body

See JsonBody

json_match_pattern

See JsonMatchPattern

label

See Label

label_match_statement

See LabelMatchStatement

label_name_condition

See LabelNameCondition

label_summary

See LabelSummary

logging_configuration

See LoggingConfiguration

logging_filter

See LoggingFilter

managed_rule_group_statement

See ManagedRuleGroupStatement

managed_rule_group_summary

See ManagedRuleGroupSummary

managed_rule_group_version

See ManagedRuleGroupVersion

managed_rule_set

See ManagedRuleSet

managed_rule_set_summary

See ManagedRuleSetSummary

managed_rule_set_version

See ManagedRuleSetVersion

method

See Method

none_action

See NoneAction

not_statement

See NotStatement

or_statement

See OrStatement

override_action

See OverrideAction

query_string

See QueryString

rate_based_statement

See RateBasedStatement

rate_based_statement_managed_keys_ip_set

See RateBasedStatementManagedKeysIpSet

regex

See Regex

regex_pattern_set

See RegexPatternSet

regex_pattern_set_reference_statement

See RegexPatternSetReferenceStatement

regex_pattern_set_summary

See RegexPatternSetSummary

rule

See Rule

rule_action

See RuleAction

rule_group

See RuleGroup

rule_group_reference_statement

See RuleGroupReferenceStatement

rule_group_summary

See RuleGroupSummary

rule_summary

See RuleSummary

sampled_http_request

See SampledHttpRequest

single_header

See SingleHeader

single_query_argument

See SingleQueryArgument

size_constraint_statement

See SizeConstraintStatement

sqli_match_statement

See SqliMatchStatement

statement

See Statement

tag

See Tag

tag_info_for_resource

See TagInfoForResource

text_transformation

See TextTransformation

time_window

See TimeWindow

uri_path

See UriPath

version_to_publish

See VersionToPublish

visibility_config

See VisibilityConfig

web_acl

See WebAcl

web_acl_summary

See WebAclSummary

xss_match_statement

See XssMatchStatement

Structs

ActionCondition

A single action condition for a Condition in a logging filter.

All

Inspect all of the elements that WAF has parsed and extracted from the web request

AllQueryArguments

All query arguments of a web request.

AllowAction

Specifies that WAF should allow the request and optionally defines additional

AndStatement

A logical rule statement used to combine other rule statements with AND logic. You provide more than one Statement within the AndStatement.

BlockAction

Specifies that WAF should block the request and optionally defines additional

Body

The body of a web request. This immediately follows the request headers.

ByteMatchStatement

A rule statement that defines a string match search for WAF to apply to web requests. The byte match statement provides the bytes to search for, the location in requests that you want WAF to search, and other settings. The bytes to search for are typically a string that corresponds with ASCII characters. In the WAF console and the developer guide, this is refered to as a string match statement.

Condition

A single match condition for a Filter.

CountAction

Specifies that WAF should count the request. Optionally defines additional custom

CustomHttpHeader

A custom header for custom request and response handling. This is used in CustomResponse and CustomRequestHandling.

CustomRequestHandling

Custom request handling behavior that inserts custom headers into a web request. You can

CustomResponse

A custom response to send to the client. You can define a custom response for rule

CustomResponseBody

The response body to use in a custom response to a web request. This is referenced by

DefaultAction

In a WebACL, this is the action that you want WAF to perform

ExcludedRule

Specifies a single rule to exclude from the rule group. Excluding a rule overrides its

FieldToMatch

The part of a web request that you want WAF to inspect. Include the single

Filter

A single logging filter, used in LoggingFilter.

FirewallManagerRuleGroup

A rule group that's defined for an Firewall Manager WAF policy.

FirewallManagerStatement

The processing guidance for an Firewall Manager rule. This is like a regular rule Statement, but it can only contain a rule group reference.

ForwardedIpConfig

The configuration for inspecting IP addresses in an HTTP header that you specify, instead of using the IP address that's reported by the web request origin. Commonly, this is the X-Forwarded-For (XFF) header, but you can specify any header name.

GeoMatchStatement

A rule statement used to identify web requests based on country of origin.

HttpHeader

Part of the response from GetSampledRequests. This is a complex type

HttpRequest

Part of the response from GetSampledRequests. This is a complex type

IpSet

Contains one or more IP addresses or blocks of IP addresses specified in Classless

IpSetForwardedIpConfig

The configuration for inspecting IP addresses in an HTTP header that you specify, instead of using the IP address that's reported by the web request origin. Commonly, this is the X-Forwarded-For (XFF) header, but you can specify any header name.

IpSetReferenceStatement

A rule statement used to detect web requests coming from particular IP addresses or address ranges. To use this, create an IPSet that specifies the addresses you want to detect, then use the ARN of that set in this statement. To create an IP set, see CreateIPSet.

IpSetSummary

High-level information about an IPSet, returned by operations like create and list. This provides information like the ID, that you can use to retrieve and manage an IPSet, and the ARN, that you provide to the IPSetReferenceStatement to use the address set in a Rule.

JsonBody

The body of a web request, inspected as JSON. The body immediately follows the request

JsonMatchPattern

The patterns to look for in the JSON body. WAF inspects the results of these

Label

A single label container. This is used as an element of a label array in multiple

LabelMatchStatement

A rule statement that defines a string match search against labels that have been added to the web request by rules that have already run in the web ACL.

LabelNameCondition

A single label name condition for a Condition in a logging

LabelSummary

List of labels used by one or more of the rules of a RuleGroup. This

LoggingConfiguration

Defines an association between Amazon Kinesis Data Firehose destinations and a web ACL

LoggingFilter

Filtering that specifies which web requests are kept in the logs and which are dropped,

ManagedRuleGroupStatement

A rule statement used to run the rules that are defined in a managed rule group. To use this, provide the vendor name and the name of the rule group in this statement. You can retrieve the required names by calling ListAvailableManagedRuleGroups.

ManagedRuleGroupSummary

High-level information about a managed rule group, returned by ListAvailableManagedRuleGroups. This provides information like the name and vendor name, that you provide when you add a ManagedRuleGroupStatement to a web ACL. Managed rule groups include Amazon Web Services Managed Rules rule groups, which are free of charge to WAF customers, and Marketplace managed rule groups, which you can subscribe to through Marketplace.

ManagedRuleGroupVersion

Describes a single version of a managed rule group.

ManagedRuleSet

A set of rules that is managed by Amazon Web Services and Marketplace sellers to provide versioned managed

ManagedRuleSetSummary

High-level information for a managed rule set.

ManagedRuleSetVersion

Information for a single version of a managed rule set.

Method

The HTTP method of a web request. The method indicates the type of operation that the

NoneAction

Specifies that WAF should do nothing. This is generally used to try out a rule

NotStatement

A logical rule statement used to negate the results of another rule statement. You provide one Statement within the NotStatement.

OrStatement

A logical rule statement used to combine other rule statements with OR logic. You provide more than one Statement within the OrStatement.

OverrideAction

The override action to apply to the rules in a rule group. Used only for rule statements that reference a rule group,

QueryString

The query string of a web request. This is the part of a URL that appears after a

RateBasedStatement

A rate-based rule tracks the rate of requests for each originating IP address, and triggers the rule action when the rate exceeds a limit that you specify on the number of requests in any 5-minute time span. You can use this to put a temporary block on requests from an IP address that is sending excessive requests.

RateBasedStatementManagedKeysIpSet

The set of IP addresses that are currently blocked for a rate-based statement.

Regex

A single regular expression. This is used in a RegexPatternSet.

RegexPatternSet

Contains one or more regular expressions.

RegexPatternSetReferenceStatement

A rule statement used to search web request components for matches with regular expressions. To use this, create a RegexPatternSet that specifies the expressions that you want to detect, then use the ARN of that set in this statement. A web request matches the pattern set rule statement if the request component matches any of the patterns in the set. To create a regex pattern set, see CreateRegexPatternSet.

RegexPatternSetSummary

High-level information about a RegexPatternSet, returned by operations like create and list. This provides information like the ID, that you can use to retrieve and manage a RegexPatternSet, and the ARN, that you provide to the RegexPatternSetReferenceStatement to use the pattern set in a Rule.

Rule

A single rule, which you can use in a WebACL or RuleGroup to identify web requests that you want to allow, block, or count.

RuleAction

The action that WAF should take on a web request when it matches a rule's

RuleGroup

A rule group defines a collection of rules to inspect and control web requests that you can use in a WebACL. When you create a rule group, you define an immutable capacity limit. If you update a rule group, you must stay within the capacity. This allows others to reuse the rule group with confidence in its capacity requirements.

RuleGroupReferenceStatement

A rule statement used to run the rules that are defined in a RuleGroup. To use this, create a rule group with your rules, then provide the ARN of the rule group in this statement.

RuleGroupSummary

High-level information about a RuleGroup, returned by operations like create and list. This provides information like the ID, that you can use to retrieve and manage a RuleGroup, and the ARN, that you provide to the RuleGroupReferenceStatement to use the rule group in a Rule.

RuleSummary

High-level information about a Rule, returned by operations like DescribeManagedRuleGroup. This provides information like the ID, that you can use to retrieve and manage a RuleGroup, and the ARN, that you provide to the RuleGroupReferenceStatement to use the rule group in a Rule.

SampledHttpRequest

Represents a single sampled web request. The response from GetSampledRequests includes a SampledHTTPRequests complex type

SingleHeader

One of the headers in a web request, identified by name, for example,

SingleQueryArgument

One query argument in a web request, identified by name, for example

SizeConstraintStatement

A rule statement that compares a number of bytes against the size of a request component, using a comparison operator, such as greater than (>) or less than (<). For example, you can use a size constraint statement to look for query strings that are longer than 100 bytes.

SqliMatchStatement

Attackers sometimes insert malicious SQL code into web requests in an effort to extract data from your database. To allow or block web requests that appear to contain malicious SQL code, create one or more SQL injection match conditions. An SQL injection match condition identifies the part of web requests, such as the URI or the query string, that you want WAF to inspect. Later in the process, when you create a web ACL, you specify whether to allow or block requests that appear to contain malicious SQL code.

Statement

The processing guidance for a Rule, used by WAF to determine

Tag

A tag associated with an Amazon Web Services resource. Tags are key:value pairs that you can use to

TagInfoForResource

The collection of tagging definitions for an Amazon Web Services resource. Tags are key:value pairs

TextTransformation

Text transformations eliminate some of the unusual formatting that attackers use in web

TimeWindow

In a GetSampledRequests request, the StartTime and

UriPath

The path component of the URI of a web request. This is the part of a web request that identifies a resource. For example, /images/daily-ad.jpg.

VersionToPublish

A version of the named managed rule group, that the rule group's vendor publishes for

VisibilityConfig

Defines and enables Amazon CloudWatch metrics and web request sample collection.

WebAcl

A web ACL defines a collection of rules to use to inspect and control web requests. Each rule has an action defined (allow, block, or count) for requests that match the statement of the rule. In the web ACL, you assign a default action to take (allow, block) for any request that does not match any of the rules. The rules in a web ACL can be a combination of the types Rule, RuleGroup, and managed rule group. You can associate a web ACL with one or more Amazon Web Services resources to protect. The resources can be an Amazon CloudFront distribution, an Amazon API Gateway REST API, an Application Load Balancer, or an AppSync GraphQL API.

WebAclSummary

High-level information about a WebACL, returned by operations like create and list. This provides information like the ID, that you can use to retrieve and manage a WebACL, and the ARN, that you provide to operations like AssociateWebACL.

XssMatchStatement

A rule statement that defines a cross-site scripting (XSS) match search for WAF to apply to web requests.

Enums

ActionValue

BodyParsingFallbackBehavior

ComparisonOperator

CountryCode

FallbackBehavior

FilterBehavior

FilterRequirement

ForwardedIpPosition

IpAddressVersion

JsonMatchScope

LabelMatchScope

ParameterExceptionField

PositionalConstraint

RateBasedStatementAggregateKeyType

ResourceType

ResponseContentType

Scope

TextTransformationType