Struct aws_sdk_wafv2::model::field_to_match::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for FieldToMatch

Implementations

impl Builder

pub fn single_header(self, input: SingleHeader) -> Self

Inspect a single header. Provide the name of the header to inspect, for example, User-Agent or Referer. This setting isn't case sensitive.

Example JSON: "SingleHeader": { "Name": "haystack" }

pub fn set_single_header(self, input: Option<SingleHeader>) -> Self

pub fn single_query_argument(self, input: SingleQueryArgument) -> Self

Inspect a single query argument. Provide the name of the query argument to inspect, such as UserName or SalesRegion. The name can be up to 30 characters long and isn't case sensitive.

This is used only to indicate the web request component for WAF to inspect, in the FieldToMatch specification.

Example JSON: "SingleQueryArgument": { "Name": "myArgument" }

pub fn set_single_query_argument(
self,
input: Option<SingleQueryArgument>
) -> Self

pub fn all_query_arguments(self, input: AllQueryArguments) -> Self

Inspect all query arguments.

pub fn set_all_query_arguments(self, input: Option<AllQueryArguments>) -> Self

pub fn uri_path(self, input: UriPath) -> Self

Inspect the request URI path. This is the part of a web request that identifies a resource, for example, /images/daily-ad.jpg.

pub fn set_uri_path(self, input: Option<UriPath>) -> Self

pub fn query_string(self, input: QueryString) -> Self

Inspect the query string. This is the part of a URL that appears after a ? character, if any.

pub fn set_query_string(self, input: Option<QueryString>) -> Self

pub fn body(self, input: Body) -> Self

Inspect the request body as plain text. The request body immediately follows the request headers. This is the part of a request that contains any additional data that you want to send to your web server as the HTTP request body, such as data from a form.

Note that only the first 8 KB (8192 bytes) of the request body are forwarded to WAF for inspection by the underlying host service. If you don't need to inspect more than 8 KB, you can guarantee that you don't allow additional bytes in by combining a statement that inspects the body of the web request, such as ByteMatchStatement or RegexPatternSetReferenceStatement, with a SizeConstraintStatement that enforces an 8 KB size limit on the body of the request. WAF doesn't support inspecting the entire contents of web requests whose bodies exceed the 8 KB limit.

pub fn set_body(self, input: Option<Body>) -> Self

pub fn method(self, input: Method) -> Self

Inspect the HTTP method. The method indicates the type of operation that the request is asking the origin to perform.

pub fn set_method(self, input: Option<Method>) -> Self

pub fn json_body(self, input: JsonBody) -> Self

Inspect the request body as JSON. The request body immediately follows the request headers. This is the part of a request that contains any additional data that you want to send to your web server as the HTTP request body, such as data from a form.

Note that only the first 8 KB (8192 bytes) of the request body are forwarded to WAF for inspection by the underlying host service. If you don't need to inspect more than 8 KB, you can guarantee that you don't allow additional bytes in by combining a statement that inspects the body of the web request, such as ByteMatchStatement or RegexPatternSetReferenceStatement, with a SizeConstraintStatement that enforces an 8 KB size limit on the body of the request. WAF doesn't support inspecting the entire contents of web requests whose bodies exceed the 8 KB limit.

pub fn set_json_body(self, input: Option<JsonBody>) -> Self

pub fn build(self) -> FieldToMatch

Consumes the builder and constructs a FieldToMatch