Struct aws_sdk_wafv2::model::statement::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Statement

Implementations

impl Builder

pub fn byte_match_statement(self, input: ByteMatchStatement) -> Self

A rule statement that defines a string match search for WAF to apply to web requests. The byte match statement provides the bytes to search for, the location in requests that you want WAF to search, and other settings. The bytes to search for are typically a string that corresponds with ASCII characters. In the WAF console and the developer guide, this is refered to as a string match statement.

pub fn set_byte_match_statement(self, input: Option<ByteMatchStatement>) -> Self

pub fn sqli_match_statement(self, input: SqliMatchStatement) -> Self

Attackers sometimes insert malicious SQL code into web requests in an effort to extract data from your database. To allow or block web requests that appear to contain malicious SQL code, create one or more SQL injection match conditions. An SQL injection match condition identifies the part of web requests, such as the URI or the query string, that you want WAF to inspect. Later in the process, when you create a web ACL, you specify whether to allow or block requests that appear to contain malicious SQL code.

pub fn set_sqli_match_statement(self, input: Option<SqliMatchStatement>) -> Self

pub fn xss_match_statement(self, input: XssMatchStatement) -> Self

A rule statement that defines a cross-site scripting (XSS) match search for WAF to apply to web requests. XSS attacks are those where the attacker uses vulnerabilities in a benign website as a vehicle to inject malicious client-site scripts into other legitimate web browsers. The XSS match statement provides the location in requests that you want WAF to search and text transformations to use on the search area before WAF searches for character sequences that are likely to be malicious strings.

pub fn set_xss_match_statement(self, input: Option<XssMatchStatement>) -> Self

pub fn size_constraint_statement(self, input: SizeConstraintStatement) -> Self

A rule statement that compares a number of bytes against the size of a request component, using a comparison operator, such as greater than (>) or less than (<). For example, you can use a size constraint statement to look for query strings that are longer than 100 bytes.

If you configure WAF to inspect the request body, WAF inspects only the first 8192 bytes (8 KB). If the request body for your web requests never exceeds 8192 bytes, you can create a size constraint condition and block requests that have a request body greater than 8192 bytes.

If you choose URI for the value of Part of the request to filter on, the slash (/) in the URI counts as one character. For example, the URI /logo.jpg is nine characters long.

pub fn set_size_constraint_statement(
self,
input: Option<SizeConstraintStatement>
) -> Self

pub fn geo_match_statement(self, input: GeoMatchStatement) -> Self

A rule statement used to identify web requests based on country of origin.

pub fn set_geo_match_statement(self, input: Option<GeoMatchStatement>) -> Self

pub fn rule_group_reference_statement(
self,
input: RuleGroupReferenceStatement
) -> Self

A rule statement used to run the rules that are defined in a RuleGroup. To use this, create a rule group with your rules, then provide the ARN of the rule group in this statement.

You cannot nest a RuleGroupReferenceStatement, for example for use inside a NotStatement or OrStatement. It can only be referenced as a top-level statement within a rule.

pub fn set_rule_group_reference_statement(
self,
input: Option<RuleGroupReferenceStatement>
) -> Self

pub fn ip_set_reference_statement(self, input: IpSetReferenceStatement) -> Self

A rule statement used to detect web requests coming from particular IP addresses or address ranges. To use this, create an IPSet that specifies the addresses you want to detect, then use the ARN of that set in this statement. To create an IP set, see CreateIPSet.

Each IP set rule statement references an IP set. You create and maintain the set independent of your rules. This allows you to use the single set in multiple rules. When you update the referenced set, WAF automatically updates all rules that reference it.

pub fn set_ip_set_reference_statement(
self,
input: Option<IpSetReferenceStatement>
) -> Self

pub fn regex_pattern_set_reference_statement(
self,
input: RegexPatternSetReferenceStatement
) -> Self

A rule statement used to search web request components for matches with regular expressions. To use this, create a RegexPatternSet that specifies the expressions that you want to detect, then use the ARN of that set in this statement. A web request matches the pattern set rule statement if the request component matches any of the patterns in the set. To create a regex pattern set, see CreateRegexPatternSet.

Each regex pattern set rule statement references a regex pattern set. You create and maintain the set independent of your rules. This allows you to use the single set in multiple rules. When you update the referenced set, WAF automatically updates all rules that reference it.

pub fn set_regex_pattern_set_reference_statement(
self,
input: Option<RegexPatternSetReferenceStatement>
) -> Self

pub fn rate_based_statement(self, input: RateBasedStatement) -> Self

A rate-based rule tracks the rate of requests for each originating IP address, and triggers the rule action when the rate exceeds a limit that you specify on the number of requests in any 5-minute time span. You can use this to put a temporary block on requests from an IP address that is sending excessive requests.

When the rule action triggers, WAF blocks additional requests from the IP address until the request rate falls below the limit.

You can optionally nest another statement inside the rate-based statement, to narrow the scope of the rule so that it only counts requests that match the nested statement. For example, based on recent requests that you have seen from an attacker, you might create a rate-based rule with a nested AND rule statement that contains the following nested statements:

An IP match statement with an IP set that specified the address 192.0.2.44.

A string match statement that searches in the User-Agent header for the string BadBot.

In this rate-based rule, you also define a rate limit. For this example, the rate limit is 1,000. Requests that meet both of the conditions in the statements are counted. If the count exceeds 1,000 requests per five minutes, the rule action triggers. Requests that do not meet both conditions are not counted towards the rate limit and are not affected by this rule.

You cannot nest a RateBasedStatement, for example for use inside a NotStatement or OrStatement. It can only be referenced as a top-level statement within a rule.

pub fn set_rate_based_statement(self, input: Option<RateBasedStatement>) -> Self

pub fn and_statement(self, input: AndStatement) -> Self

A logical rule statement used to combine other rule statements with AND logic. You provide more than one Statement within the AndStatement.

pub fn set_and_statement(self, input: Option<AndStatement>) -> Self

pub fn or_statement(self, input: OrStatement) -> Self

A logical rule statement used to combine other rule statements with OR logic. You provide more than one Statement within the OrStatement.

pub fn set_or_statement(self, input: Option<OrStatement>) -> Self

pub fn not_statement(self, input: NotStatement) -> Self

A logical rule statement used to negate the results of another rule statement. You provide one Statement within the NotStatement.

pub fn set_not_statement(self, input: Option<NotStatement>) -> Self

pub fn managed_rule_group_statement(
self,
input: ManagedRuleGroupStatement
) -> Self

A rule statement used to run the rules that are defined in a managed rule group. To use this, provide the vendor name and the name of the rule group in this statement. You can retrieve the required names by calling ListAvailableManagedRuleGroups.

You cannot nest a ManagedRuleGroupStatement, for example for use inside a NotStatement or OrStatement. It can only be referenced as a top-level statement within a rule.

pub fn set_managed_rule_group_statement(
self,
input: Option<ManagedRuleGroupStatement>
) -> Self

pub fn label_match_statement(self, input: LabelMatchStatement) -> Self

A rule statement that defines a string match search against labels that have been added to the web request by rules that have already run in the web ACL.

The label match statement provides the label or namespace string to search for. The label string can represent a part or all of the fully qualified label name that had been added to the web request. Fully qualified labels have a prefix, optional namespaces, and label name. The prefix identifies the rule group or web ACL context of the rule that added the label. If you do not provide the fully qualified name in your label match string, WAF performs the search for labels that were added in the same context as the label match statement.

pub fn set_label_match_statement(
self,
input: Option<LabelMatchStatement>
) -> Self

pub fn build(self) -> Statement

Consumes the builder and constructs a Statement