Struct aws_sdk_wafv2::model::WebAcl

#[non_exhaustive]
pub struct WebAcl {

    pub name: Option<String>,
    pub id: Option<String>,
    pub arn: Option<String>,
    pub default_action: Option<DefaultAction>,
    pub description: Option<String>,
    pub rules: Option<Vec<Rule>>,
    pub visibility_config: Option<VisibilityConfig>,
    pub capacity: i64,
    pub pre_process_firewall_manager_rule_groups: Option<Vec<FirewallManagerRuleGroup>>,
    pub post_process_firewall_manager_rule_groups: Option<Vec<FirewallManagerRuleGroup>>,
    pub managed_by_firewall_manager: bool,
    pub label_namespace: Option<String>,
    pub custom_response_bodies: Option<HashMap<String, CustomResponseBody>>,

}

A web ACL defines a collection of rules to use to inspect and control web requests. Each rule has an action defined (allow, block, or count) for requests that match the statement of the rule. In the web ACL, you assign a default action to take (allow, block) for any request that does not match any of the rules. The rules in a web ACL can be a combination of the types Rule, RuleGroup, and managed rule group. You can associate a web ACL with one or more Amazon Web Services resources to protect. The resources can be an Amazon CloudFront distribution, an Amazon API Gateway REST API, an Application Load Balancer, or an AppSync GraphQL API.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The name of the web ACL. You cannot change the name of a web ACL after you create it.

id: Option<String>

A unique identifier for the WebACL. This ID is returned in the responses to create and list commands. You use this ID to do things like get, update, and delete a WebACL.

arn: Option<String>

The Amazon Resource Name (ARN) of the web ACL that you want to associate with the resource.

default_action: Option<DefaultAction>

The action to perform if none of the Rules contained in the WebACL match.

description: Option<String>

A description of the web ACL that helps with identification.

rules: Option<Vec<Rule>>

The Rule statements used to identify the web requests that you want to allow, block, or count. Each rule includes one top-level statement that WAF uses to identify matching web requests, and parameters that govern how WAF handles them.

visibility_config: Option<VisibilityConfig>

Defines and enables Amazon CloudWatch metrics and web request sample collection.

capacity: i64

The web ACL capacity units (WCUs) currently being used by this web ACL.

WAF uses WCUs to calculate and control the operating resources that are used to run your rules, rule groups, and web ACLs. WAF calculates capacity differently for each rule type, to reflect the relative cost of each rule. Simple rules that cost little to run use fewer WCUs than more complex rules that use more processing power. Rule group capacity is fixed at creation, which helps users plan their web ACL WCU usage when they use a rule group. The WCU limit for web ACLs is 1,500.

pre_process_firewall_manager_rule_groups: Option<Vec<FirewallManagerRuleGroup>>

The first set of rules for WAF to process in the web ACL. This is defined in an Firewall Manager WAF policy and contains only rule group references. You can't alter these. Any rules and rule groups that you define for the web ACL are prioritized after these.

In the Firewall Manager WAF policy, the Firewall Manager administrator can define a set of rule groups to run first in the web ACL and a set of rule groups to run last. Within each set, the administrator prioritizes the rule groups, to determine their relative processing order.

post_process_firewall_manager_rule_groups: Option<Vec<FirewallManagerRuleGroup>>

The last set of rules for WAF to process in the web ACL. This is defined in an Firewall Manager WAF policy and contains only rule group references. You can't alter these. Any rules and rule groups that you define for the web ACL are prioritized before these.

In the Firewall Manager WAF policy, the Firewall Manager administrator can define a set of rule groups to run first in the web ACL and a set of rule groups to run last. Within each set, the administrator prioritizes the rule groups, to determine their relative processing order.

managed_by_firewall_manager: bool

Indicates whether this web ACL is managed by Firewall Manager. If true, then only Firewall Manager can delete the web ACL or any Firewall Manager rule groups in the web ACL.

label_namespace: Option<String>

The label namespace prefix for this web ACL. All labels added by rules in this web ACL have this prefix.

The syntax for the label namespace prefix for a web ACL is the following:

awswaf::webacl::

When a rule with a label matches a web request, WAF adds the fully qualified label to the request. A fully qualified label is made up of the label namespace from the rule group or web ACL where the rule is defined and the label from the rule, separated by a colon:

:

custom_response_bodies: Option<HashMap<String, CustomResponseBody>>

A map of custom response keys and content bodies. When you create a rule with a block action, you can send a custom response to the web request. You define these for the web ACL, and then use them in the rules and default actions that you define in the web ACL.

For information about customizing web requests and responses, see Customizing web requests and responses in WAF in the WAF Developer Guide.

For information about the limits on count and size for custom request and response settings, see WAF quotas in the WAF Developer Guide.

Implementations

impl WebAcl

pub fn builder() -> Builder

Creates a new builder-style object to manufacture WebAcl