Module aws_sdk_wafv2::operation

Structs

AssociateWebACL

Associates a web ACL with a regional application resource, to protect the resource.

CheckCapacity

Returns the web ACL capacity unit (WCU) requirements for a specified scope and set of rules.

CreateIPSet

Creates an IPSet, which you use to identify web requests that

CreateRegexPatternSet

Creates a RegexPatternSet, which you reference in a RegexPatternSetReferenceStatement, to have WAF inspect a web request

CreateRuleGroup

Creates a RuleGroup per the specifications provided.

CreateWebACL

Creates a WebACL per the specifications provided.

DeleteFirewallManagerRuleGroups

Deletes all rule groups that are managed by Firewall Manager for the specified web ACL.

DeleteIPSet

Deletes the specified IPSet.

DeleteLoggingConfiguration

Deletes the LoggingConfiguration from the specified web ACL.

DeletePermissionPolicy

Permanently deletes an IAM policy from the specified rule group.

DeleteRegexPatternSet

Deletes the specified RegexPatternSet.

DeleteRuleGroup

Deletes the specified RuleGroup.

DeleteWebACL

Deletes the specified WebACL.

DescribeManagedRuleGroup

Provides high-level information for a managed rule group, including descriptions of the

DisassociateWebACL

Disassociates a web ACL from a regional application resource. A regional application can be an Application Load Balancer (ALB), an Amazon API Gateway REST API, or an AppSync GraphQL API.

GetIPSet

Retrieves the specified IPSet.

GetLoggingConfiguration

Returns the LoggingConfiguration for the specified web ACL.

GetManagedRuleSet

Retrieves the specified managed rule set.

GetPermissionPolicy

Returns the IAM policy that is attached to the specified rule group.

GetRateBasedStatementManagedKeys

Retrieves the keys that are currently blocked by a rate-based rule. The maximum number

GetRegexPatternSet

Retrieves the specified RegexPatternSet.

GetRuleGroup

Retrieves the specified RuleGroup.

GetSampledRequests

Gets detailed information about a specified number of requests--a sample--that WAF

GetWebACL

Retrieves the specified WebACL.

GetWebACLForResource

Retrieves the WebACL for the specified resource.

ListAvailableManagedRuleGroupVersions

Returns a list of the available versions for the specified managed rule group.

ListAvailableManagedRuleGroups

Retrieves an array of managed rule groups that are available for you to use. This list

ListIPSets

Retrieves an array of IPSetSummary objects for the IP sets that you

ListLoggingConfigurations

Retrieves an array of your LoggingConfiguration objects.

ListManagedRuleSets

Retrieves the managed rule sets that you own.

ListRegexPatternSets

Retrieves an array of RegexPatternSetSummary objects for the regex

ListResourcesForWebACL

Retrieves an array of the Amazon Resource Names (ARNs) for the regional resources that

ListRuleGroups

Retrieves an array of RuleGroupSummary objects for the rule groups

ListTagsForResource

Retrieves the TagInfoForResource for the specified resource. Tags are

ListWebACLs

Retrieves an array of WebACLSummary objects for the web ACLs that you

PutLoggingConfiguration

Enables the specified LoggingConfiguration, to start logging from a

PutManagedRuleSetVersions

Defines the versions of your managed rule set that you are offering to the customers.

PutPermissionPolicy

Attaches an IAM policy to the specified resource. Use this to share a rule group across

TagResource

Associates tags with the specified Amazon Web Services resource. Tags are key:value pairs that you can

UntagResource

Disassociates tags from an Amazon Web Services resource. Tags are key:value pairs that you can

UpdateIPSet

Updates the specified IPSet.

UpdateManagedRuleSetVersionExpiryDate

Updates the expiration information for your managed rule set. Use this to initiate the

UpdateRegexPatternSet

Updates the specified RegexPatternSet.

UpdateRuleGroup

Updates the specified RuleGroup.

UpdateWebACL

Updates the specified WebACL.